requirements.txt
streamlit
google-cloud-firestore
google-auth-oauthlib
plotly